Moving time converter to shared utils
